Handle a presumed API change
The previous code suggests that the API response when querying a current DNS record that isn't set would be "0.0.0.0", but my responses are coming back with the name of one of NFS's name servers. It's important to catch this case because a failed attempt to remove a DNS record crashes out of the script entirely.